Add files via upload
This is the last update of the PRIO Battle Deaths Dataset (v.1.0 and v.2.0) first published in 2005 and updated in 2006. The dataset is compatible with the UCDP/PRIO Armed Conflict Dataset (Gleditsch et al., 2002; Harbom & Wallensteen, 2009) for the period 1946–2008. Versions 1.0 and 2.0 are also compatible with the Correlates of War Dataset (Sarkees, 2000) for the period 1900–97; and the Fearon & Laitin (2003) dataset on civil conflict in the period 1945–99. Since the Correlates of War and Fearon & Laitin datasets have not been updated, we refer to Version 2.0 for data and documentation relating to these two datasets.